+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Introduce `AnnotationsUtil.getRepeatingAnnotations(...)` utility.
+
 ### [v0.01](https://github.com/realityforge/proton-processor-pack/tree/v0.01) (2019-12-24) · [Full Changelog](https://github.com/realityforge/proton-processor-pack/compare/5d8d0136c796a3732c5d74715aa5e01764a9eaa9...v0.01)
 
 Changes in this release: